backend/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.2
Django==4.2
django-cors-headers==4.2.0
django-qrcode==0.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
Pillow==10.0.0
PyJWT==1.7.1
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rest-framework-simplejwt==0.0.2
segno==1.5.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
